Create Codes that I've inspired or practiced
This code sample is from one of the ticketing requirements: description and response. It took me a lot of time to understand and practise because I need to understand Object-Oriented Programming profoundly (OOP), or else I might have difficulty creating code and become messy. 

So I practice creating a blueprint for creating objects ( Using Classes) and then creating codes using if/else statements, counts, dictionaries, and append.

Also, it inspires me to learn more about using the OOP, especially classes and objects. In the future, I want to learn this OOP further.